smithing to smelting & smitheditem to smithing
