+++ src/tasks/minions/smithingActivity.ts
@@ -2,48 +2,28 @@ import { Task, KlasaMessage } from 'klasa';
 
 import { saidYes, noOp } from '../../lib/util';
 import { Time } from '../../lib/constants';
+import { SkillsEnum } from '../../lib/skilling/types';
 import { SmithingActivityTaskOptions } from '../../lib/types/minions';
 import getUsersPerkTier from '../../lib/util/getUsersPerkTier';
-import Smithing from '../../lib/skilling/skills/smithing';
-import { rand } from 'oldschooljs/dist/util/util';
 import { channelIsSendable } from '../../lib/util/channelIsSendable';
-import itemID from '../../lib/util/itemID';
-import { SkillsEnum } from '../../lib/skilling/types';
+import Smithing from '../../lib/skilling/skills/smithing';
 
 export default class extends Task {
-	async run({ barID, quantity, userID, channelID }: SmithingActivityTaskOptions) {
+	async run({ smithedBarID, quantity, userID, channelID }: SmithingActivityTaskOptions) {
 		const user = await this.client.users.fetch(userID);
 		const currentLevel = user.skillLevel(SkillsEnum.Smithing);
 
-		const bar = Smithing.Bars.find(bar => bar.id === barID);
-		if (!bar) return;
+		const SmithedBar = Smithing.SmithedBars.find(SmithedBar => SmithedBar.id === smithedBarID);
+		if (!SmithedBar) return;
 
-		// If this bar has a chance of failing to smith, calculate that here.
-		const oldQuantity = quantity;
-		if (bar.chanceOfFail > 0) {
-			let newQuantity = 0;
-			for (let i = 0; i < quantity; i++) {
-				if (rand(0, 100) < bar.chanceOfFail) {
-					newQuantity++;
-				}
-			}
-			quantity = newQuantity;
-		}
-
-		let xpReceived = quantity * bar.xp;
-
-		if (
-			bar.id === itemID('Gold bar') &&
-			user.hasItemEquippedAnywhere(itemID('Goldsmith gauntlets'))
-		) {
-			xpReceived = quantity * 56.2;
-		}
+		const xpReceived = quantity * SmithedBar.xp;
 
 		await user.addXP(SkillsEnum.Smithing, xpReceived);
 		const newLevel = user.skillLevel(SkillsEnum.Smithing);
 
-		let str = `${user}, ${user.minionName} finished smithing ${quantity}x ${
-			bar.name
+		let str = `${user}, ${user.minionName} finished smithing ${quantity *
+			SmithedBar.outputMultiple}x ${
+			SmithedBar.name
 		}, you also received ${xpReceived.toLocaleString()} XP. ${
 			user.minionName
 		} asks if you'd like them to do another of the same trip.`;
@@ -52,12 +32,8 @@ export default class extends Task {
 			str += `\n\n${user.minionName}'s Smithing level is now ${newLevel}!`;
 		}
 
-		if (bar.chanceOfFail > 0 && oldQuantity > quantity) {
-			str += `\n\n${oldQuantity - quantity} ${bar.name}s failed to smelt.`;
-		}
-
 		const loot = {
-			[bar.id]: quantity
+			[SmithedBar.id]: quantity * SmithedBar.outputMultiple
 		};
 
 		await user.addItemsToBank(loot, true);
@@ -78,11 +54,11 @@ export default class extends Task {
 				if (response) {
 					if (response.author.minionIsBusy) return;
 
-					user.log(`continued trip of ${oldQuantity}x ${bar.name}[${bar.id}]`);
+					user.log(`continued trip of  ${SmithedBar.name}[${SmithedBar.id}]`);
 
 					this.client.commands
-						.get('smelt')!
-						.run(response as KlasaMessage, [oldQuantity, bar.name]);
+						.get('smith')!
+						.run(response as KlasaMessage, [SmithedBar.name]);
 				}
 			})
 			.catch(noOp);
